increased int because it might get out of range

+++ src/folder/folder.entity.ts
@@ -18,6 +18,6 @@ export class Folder {
     @Exclude()
     @DeleteDateColumn({ name: 'deleted_at' })
     deletedAt: Date;
-    @Column({array:true,type:'int',default:[]})
+    @Column({array:true,type:'bigint',default:[]})
     commits: number[];
 }
